+++ src/middleware.ts
@@ -1,14 +1,7 @@
-import { auth } from '@/auth'
-import { NextResponse } from 'next/server'
+import NextAuth from 'next-auth'
+import { authConfig } from '@/auth.config'
 
-export default auth((req) => {
-  if (!req.auth && req.nextUrl.pathname.startsWith('/recipes')) {
-    const url = new URL('/api/auth/signin', req.nextUrl.origin)
-    url.searchParams.set('callbackUrl', `${req.nextUrl.pathname}${req.nextUrl.search}`)
-    return NextResponse.redirect(url)
-  }
-  return NextResponse.next()
-})
+export const { auth: middleware } = NextAuth(authConfig)
 
 export const config = {
   matcher: ['/recipes/:path*'],